Update 03.03.23
New animations implemented.

Now the player is being able to control or change the control  to the robot but we still need to implement it alongside with using the computer, making the player stay un using computer animation and controller the robot and a way to bring the control to the main player after.

more to come...