Update 10.03.23
Improved the first scene.

Add one more robot that will only work in the first scene.

Still need to add a tutorial about how to walk, jump and interact with the computer.

still need to add a key to show the player the computer button and the rest of the interactions.

 a a door for the robot to open maybe woul be nice,

more to come...