Update 21.03.23.2
Improved spring scene.

Added transition in the spring scene.

Adjusted the player jump and gravity.

Still need to adjust the camera to fir in the project but thats for a close future.

still need to decorate the spring level.

Need to make a teleporter that checks if all the letters have been collected.

more to come...