Update 23.02.23.2
Updated the spring scene now it has a tries couter, everytime the player falls in the void he will respawn in the beggining but wo'nt lose anything, if it gets to zero he will start the level over by losing his progress in the level.

still need to make the player blink when he respawns maybe using InvokeRepeating need to check it out.

needs to make the game over it wo'nt kill the player it will just restart the level.

more to come...