Update 24.03.23
Improved Player gameplay and spring scene now the player flashes when takes dmg and once he loses all the tries he is obligated to start the level all over from the beggining.

We still need a GameOver or something related to it, we gonna check it out.

some more props were added to decorate the level.

more to come...