Update 28.02.23
Game Over Screen is done we might need to find another background for it but for now it works just fine as I said.

Added a group os crowns to make the level more alive with living animals with their own behavior and routine but i'm not sure if we gonna keep them, we need to check it out.

more to come...