Update 03.04.23.2
Finally fixed the camera siuze to every single level YES!

Added the key objet of the summer level wich is a ice cream.

we are still not sure if we gonna add enemies or sth.

Now we just need to finish some details of the levels and thats it.

more to come...