Update 13.04.23
Improved the autumn level A LOT!

Almost finished already to be honest, just need details and the other minor things like places for the player to die and everything, you know what i mean.

still need to make the logic of the letter tags in the letter controller script.

more to come...