Update 27.04.23
Boss if finally finished.

Added a new camera only for the robot, still need to make it work in the previous levels, including tutorial.

Still need to finish the autumn level.
We need to check if there's any bugs in previous levels because of the scriptable objects, specially gameover screen and letter canvas related.

more to come...