Update 09.05.23.2
Improved Winter level.

Enemy fully implemented now it spawn from broken ice and destroys or can be defeated by the robot.

once the robot loses all of its 3 tries the player loses 1, if the player loses all the game over mechanic works as in other levels.

Just need to implement the Boss and add details such as snow in the screen with particles and some other cool things to make the winter level more beautiful.

more to come...