Update 10.05.23
Implemented the last BOSS, it is a Snow Golem that throws giant snow balls towards the player/Tanker Robot.

The Snow Ball already do damage on the player .

The boss can be defeated however still need some special effects for being defeated.

The attack state needs to be implemented in the state machine scripts on animator using the fire point to instantiate the ball only when it attacks and needs to be attacking eveyr once in a while.

Still need to make a barier that can go up and down or something to protect the player.

GOD is ahead of our project, nothing can go wrong.

more to come...