Update 11.05.23
Winter level final boss almost finished.

Still need to adjust the waiting time but it will work just fine for sure.

Check if the animation and the throwing time actually match.

After that the scene logic is finished and we just need ot focus on details to make the level beautiful and the transition black screen.

GOD is good.

more to come...