Update 12.05.23
Some minor improvments.

Now the player can see where he or she needs to stand to not take any dmg.

still need to make the logic for the barrier to dissapear once the boss is defeated, need to chek if its going to be in boss logic scene script or in boxx script.

more to come...