Update 31.05.23
Add a menu scene.

Need to record a video of the scientist (we haven't decided on a nome for him yet) in the lab scene the one right after the tutorial with him like using the machines in the scenario and eveyrthing to make it as the background video of the menu.

Thank you GOD!

more to come...