Update 11.07.23
The visual part of the Main menu is finished.

Start button is working just fine.

Still need to make a "Exit" button.

Still need to fix the trasition of the summer level.

we'll try to create something in between every level to create an ceirtain story of the game, maybe something informative about the season/key object or something relate

More to come..

Even though I walk through the valley of the shadow of death, I will fear no evil, for You are with me; Your rod and Your staff, they comfort me.
Psalm23:4