Update 18.07.23
Improved main menu animation.

Fixed the transition in the end of the summer level, the transition when you finish the level.

Still need to add something, some visual style to show the player that the bear boss took some damage.

Still need to add something between the levels to indicate a kind of little story or something like this.

"Love is patient and kind; love does not envy or boast; it is not arrogant or rude. It does not insist on its own way; it is not irritable or resentful; it does not rejoice at wrongdoing, but rejoices with the truth."

1 Corinthians 13:4–8.