Update 25.07.23
Started to make the information screens after every level, still finishing the first one in the spring and than just make the same in the other ones.

We still need to make an end screen and i'm sure God will help us out.

"You shall love the Lord your God with all your heart, and with all your soul, and with all your mind, and with all your strength. "
Mark 12:30