Update 28.02.23
Information scene of the spring level is almost complete, just need to finish it and go to the lab.

Once its done i need to change the teleport in the scenes to it respective information scenes.

"But blessed is the one who trusts in the LORD, whose confidence is in him. "
Jeremiah 17:7