Update 29.07.23
Information scenes of Spring and Summer are done as well as their respective levels teleporters sending the player to those screens.

Still need to make the same for Autumn and Winter.

Still need to create and end screen.

"For to me to live is Christ, and to die is gain"
Phillipians 1:21