Update 01.08.23
All info scenes are finished and workig how it is supposed to.

Still need to create an "END" scene and remake the "GameOver" Panel.

Thanks GOD for everything that we've done so far and everything we will still do in our lives. Amen.

"Do not conform to the pattern of this world, but be transformed by the renewing of your mind. Then you will be able to test and approve what God’s will is—his good, pleasing and perfect will."
Romans 12:2