Update 01.09.23
Chronometer almost done.

Still need to make it work properly and fix the scene controller method for the posgamestart.

We still need to add the informations between scenes and a final scene for it to show the metrics like time it took and to activate the pos game.

God has everything under control.