Update 27.09.23
Galerry working  for half of the drawnings we have , strill need to decide how to unlock the rest of them.

Some bugs such as the red robot getting up in the middle of a tunnel are fixed.

Still trying to fix the wall jumping boost bug.

Still need to make hte loic where it changes the image in the last scene of the pos game to work and decide how players will get them.

If we have some sparing time we will try to add a save game and sounds to the game.

GOD IS GOOD.